+++ linstor_client/commands/storpool_cmds.py
@@ -49,10 +49,6 @@ class StoragePoolCommands(Commands):
         LONG = "remotespdk"
         SHORT = "remotespdk"
 
-    class Exos(object):
-        LONG = "exos"
-        SHORT = "exos"
-
     class StorageSpaces(object):
         LONG = "storagespaces"
         SHORT = "storagespaces"
@@ -133,7 +129,6 @@ class StoragePoolCommands(Commands):
             StoragePoolCommands.FileThin,
             StoragePoolCommands.SPDK,
             StoragePoolCommands.RemoteSPDK,
-            StoragePoolCommands.Exos,
             StoragePoolCommands.StorageSpaces,
             StoragePoolCommands.StorageSpacesThin,
             StoragePoolCommands.EbsInit
@@ -296,24 +291,6 @@ class StoragePoolCommands(Commands):
         )
         p_new_file_thin_pool.set_defaults(func=self.create, driver=linstor.StoragePoolDriver.FILEThin)
 
-        p_new_exos_pool = create_subp.add_parser(
-            StoragePoolCommands.Exos.LONG,
-            aliases=[StoragePoolCommands.Exos.SHORT],
-            description='Create an EXOS storage pool'
-        )
-        self._create_pool_args(p_new_exos_pool)
-        p_new_exos_pool.add_argument(
-            'enclosure_name',
-            type=str,
-            help='Enclosure name'
-        )
-        p_new_exos_pool.add_argument(
-            'pool_sn',
-            type=str,
-            help='Exos Pool Serial Number'
-        )
-        p_new_exos_pool.set_defaults(func=self.create_exos, driver=linstor.StoragePoolDriver.EXOS)
-
         p_new_ebs_init_pool = create_subp.add_parser(
             StoragePoolCommands.EbsInit.LONG,
             aliases=[StoragePoolCommands.EbsInit.SHORT],
@@ -426,24 +403,6 @@ class StoragePoolCommands(Commands):
                 args.driver_pool_name,
                 shared_space=shrd_space,
                 external_locking=ext_locking
-            )
-        except linstor.LinstorError as e:
-            raise ArgumentError(e.message)
-        return self.handle_replies(args, replies)
-
-    def create_exos(self, args):
-        try:
-            # no shared-space and no external locking. shared-space calculated by server, external locking not allowed
-            props = {
-                apiconsts.NAMESPC_EXOS + '/' + apiconsts.KEY_STOR_POOL_EXOS_ENCLOSURE: args.enclosure_name,
-                apiconsts.NAMESPC_EXOS + '/' + apiconsts.KEY_STOR_POOL_EXOS_POOL_SN: args.pool_sn
-            }
-            replies = self.get_linstorapi().storage_pool_create(
-                args.node_name,
-                args.name,
-                args.driver,
-                args.enclosure_name + '_' + args.pool_sn,
-                property_dict=props
             )
         except linstor.LinstorError as e:
             raise ArgumentError(e.message)