+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - snapshot list and set property commands
 
+### Fixed
+
+- Fixed incorrect unit in out of range error message
+
 ## [1.25.4] - 2025-04-10
 
 ### Fixed